// src/types/models/BasketItemsResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ProductResponse } from './ProductResponse';
export type BasketItemsResponse = {
    id: number;
    basket_id?: number;
    product: ProductResponse;
    quantity: number;
    price_id_by_the_user: number;
};